requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
django==4.2.3
    # via
    #   -r requirements.in
    #   django-versatileimagefield
    #   djangorestframework
django-versatileimagefield==3.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
pillow==9.5.0
    # via
    #   -r requirements.in
    #   django-versatileimagefield
python-magic==0.4.27
    # via django-versatileimagefield
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via asgiref
